fix: separate component registration from (client) component definitions
